requirements.txt
Django>=3.1,<3.2
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
django-environ
requests-oauthlib
oauthlib
pyjwt
markdown
pylint-django
gunicorn
whitenoise
cryptography
redis
celery
django-extensions